Add edit_entry function view
This view firstly checks the existence of the entry to edit. If
the request is for non-existent entry, it redirects user to not found
page.
Then, this view proceeds according to the request method. For POST
it binds the form with POST data, validates it, and if validation
successful, saves the entry and redirects the user to its page.
If validation was not successful, the view returns populated form
back to the user.
For GET (& other methods) it binds the form with existing data, and
passes the form to context.