Add view responisble for creating new entry
This view starts with the flag already_exists, initially False.
For the POST method it binds the form with POST data, checks the
existence of entry, and if the flag is unchanged, saves the entry.
For the GET method, it creates new form instance and passes it to the
context.